# swagger/swagger.json
{
  "swagger": "2.0",
  "paths": {
...
    "/fcc/config/pre_publish/": {
      "post": {
        "description": "Get group list",
        "tags": [
          "Fcc"
        ],
        "operationId": "CommonReq",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/FccPrePublishReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/BaseResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/FetchGroupsRet"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/fcc/config/publish/": {
      "post": {
        "description": "publish config",
        "tags": [
          "Fcc"
        ],
        "operationId": "CommonReq",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/FccPublishReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/BaseResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {}
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/fcc/config/save/": {
      "post": {
        "description": "Get group list",
        "tags": [
          "Fcc"
        ],
        "operationId": "CommonReq",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/FccSaveConfigReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/BaseResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {}
                  }
                }
              ]
            }
          }
        }
      }
    },
...
    "/fcc/group/save/": {
      "post": {
        "description": "Get group list",
        "tags": [
          "Fcc"
        ],
        "operationId": "CommonReq",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/FccSaveGroupReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/BaseResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {}
                  }
                }
              ]
            }
          }
        }
      }
    },
...
    "/fcc/project/save/": {
      "post": {
        "description": "Get group list",
        "tags": [
          "Fcc"
        ],
        "operationId": "CommonReq",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/FccSaveProjectReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/BaseResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/FetchGroupsRet"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/starling/group/list/": {
      "get": {
        "description": "Get group list",
        "tags": [
          "Starling"
        ],
        "operationId": "StarlingGroupListReq",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Limit",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Offset",
            "name": "offset",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "ProjectKey",
            "name": "project_key",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "GroupKey",
            "name": "group_key",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "GroupName",
            "name": "group_name",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/BaseResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/FetchGroupsRet"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/starling/group/save/": {
      "post": {
        "description": "Get group list",
        "tags": [
          "Starling"
        ],
        "operationId": "CommonReq",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/StarlingSaveGroupReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/BaseResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {}
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/starling/origin/list/": {
      "get": {
        "description": "Get config list",
        "tags": [
          "Starling"
        ],
        "operationId": "StarlingOriginLgListReq",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Limit",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Offset",
            "name": "offset",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "ProjectKey",
            "name": "project_key",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "GroupKey",
            "name": "group_key",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "LangKey",
            "name": "conf_key",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/BaseResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/FetchOriginLgsRet"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/starling/origin/save/": {
      "post": {
        "description": "Get group list",
        "tags": [
          "Starling"
        ],
        "operationId": "CommonReq",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/StarlingSaveOriginLgReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/BaseResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {}
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/starling/project/list/": {
      "get": {
        "description": "Get project list",
        "tags": [
          "Starling"
        ],
        "operationId": "StarlingProjectListReq",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Limit",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Offset",
            "name": "offset",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "ProjectKey",
            "name": "project_key",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "ProjectName",
            "name": "project_name",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/BaseResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/FetchProjectsRet"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/starling/project/save/": {
      "post": {
        "description": "Get group list",
        "tags": [
          "Starling"
        ],
        "operationId": "CommonReq",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/StarlingSaveProjectReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/BaseResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/FetchGroupsRet"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/starling/translate/save/": {
      "post": {
        "description": "Get group list",
        "tags": [
          "Starling"
        ],
        "operationId": "CommonReq",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/StarlingSaveTransLgReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/BaseResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {}
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/starling/translation/": {
      "get": {
        "description": "Get project list",
        "tags": [
          "Starling"
        ],
        "operationId": "StarlingGetTransLgReq",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ProjectKey",
            "name": "project_key",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "GroupKey",
            "name": "group_key",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "LangKey",
            "name": "lang_key",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "Lang",
            "name": "lang",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/BaseResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/TransLg"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/starling/translation/list/": {
      "get": {
        "description": "Get config list",
        "tags": [
          "Starling"
        ],
        "operationId": "StarlingTransLgListReq",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Limit",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Offset",
            "name": "offset",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "ProjectKey",
            "name": "project_key",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "GroupKey",
            "name": "group_key",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "LangKey",
            "name": "lang_key",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/BaseResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/FetchTransLgsRet"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    }
  },
  "definitions": {
    "BaseResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Code"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        }
      },
      "x-go-package": "github.com/qudj/fly_api/responses"
    },
...
    "FccPrePublishReq": {
      "type": "object",
      "properties": {
        "conf_key": {
          "type": "string",
          "x-go-name": "ConfKey"
        },
        "group_key": {
          "type": "string",
          "x-go-name": "GroupKey"
        },
        "pre_value": {
          "type": "string",
          "x-go-name": "PreValue"
        },
        "project_key": {
          "type": "string",
          "x-go-name": "ProjectKey"
        }
      },
      "x-go-package": "github.com/qudj/fly_api/models"
    },
    "FccPublishReq": {
      "type": "object",
      "properties": {
        "conf_key": {
          "type": "string",
          "x-go-name": "ConfKey"
        },
        "group_key": {
          "type": "string",
          "x-go-name": "GroupKey"
        },
        "project_key": {
          "type": "string",
          "x-go-name": "ProjectKey"
        }
      },
      "x-go-package": "github.com/qudj/fly_api/models"
    },
    "FccSaveConfigReq": {
      "type": "object",
      "properties": {
        "conf_key": {
          "type": "string",
          "x-go-name": "ConfKey"
        },
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "group_key": {
          "type": "string",
          "x-go-name": "GroupKey"
        },
        "project_key": {
          "type": "string",
          "x-go-name": "ProjectKey"
        },
        "save_mode": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "SaveMode"
        },
        "status": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Status"
        }
      },
      "x-go-package": "github.com/qudj/fly_api/models"
    },
    "FccSaveGroupReq": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "group_key": {
          "type": "string",
          "x-go-name": "GroupKey"
        },
        "group_name": {
          "type": "string",
          "x-go-name": "GroupName"
        },
        "project_key": {
          "type": "string",
          "x-go-name": "ProjectKey"
        },
        "save_mode": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "SaveMode"
        },
        "status": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Status"
        }
      },
      "x-go-package": "github.com/qudj/fly_api/models"
    },
    "FccSaveProjectReq": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "project_key": {
          "type": "string",
          "x-go-name": "ProjectKey"
        },
        "project_name": {
          "type": "string",
          "x-go-name": "ProjectName"
        },
        "save_mode": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "SaveMode"
        },
        "status": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Status"
        }
      },
      "x-go-package": "github.com/qudj/fly_api/models"
    },
...
    "FetchGroupsRet": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Group"
          },
          "x-go-name": "List"
        },
        "total": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Total"
        }
      },
      "x-go-package": "github.com/qudj/fly_lib/models/proto/fly_starling_serv"
    },
...
    "FetchOriginLgsRet": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OriginLg"
          },
          "x-go-name": "List"
        },
        "total": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Total"
        }
      },
      "x-go-package": "github.com/qudj/fly_lib/models/proto/fly_starling_serv"
    },
...
    "FetchProjectsRet": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Project"
          },
          "x-go-name": "List"
        },
        "total": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Total"
        }
      },
      "x-go-package": "github.com/qudj/fly_lib/models/proto/fcc_serv"
    },
...
    "FetchTransLgsRet": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TransLg"
          },
          "x-go-name": "List"
        },
        "total": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Total"
        }
      },
      "x-go-package": "github.com/qudj/fly_lib/models/proto/fly_starling_serv"
    },
...
    "StarlingSaveGroupReq": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "group_key": {
          "type": "string",
          "x-go-name": "GroupKey"
        },
        "group_name": {
          "type": "string",
          "x-go-name": "GroupName"
        },
        "project_key": {
          "type": "string",
          "x-go-name": "ProjectKey"
        },
        "save_mode": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "SaveMode"
        },
        "status": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Status"
        }
      },
      "x-go-package": "github.com/qudj/fly_api/models"
    },
    "StarlingSaveOriginLgReq": {
      "type": "object",
      "properties": {
        "group_key": {
          "type": "string",
          "x-go-name": "GroupKey"
        },
        "lang": {
          "type": "string",
          "x-go-name": "Lang"
        },
        "lang_key": {
          "type": "string",
          "x-go-name": "LangKey"
        },
        "origin_text": {
          "type": "string",
          "x-go-name": "OriginText"
        },
        "project_key": {
          "type": "string",
          "x-go-name": "ProjectKey"
        },
        "save_mode": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "SaveMode"
        },
        "status": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Status"
        }
      },
      "x-go-package": "github.com/qudj/fly_api/models"
    },
    "StarlingSaveProjectReq": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "project_key": {
          "type": "string",
          "x-go-name": "ProjectKey"
        },
        "project_name": {
          "type": "string",
          "x-go-name": "ProjectName"
        },
        "save_mode": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "SaveMode"
        },
        "status": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Status"
        }
      },
      "x-go-package": "github.com/qudj/fly_api/models"
    },
    "StarlingSaveTransLgReq": {
      "type": "object",
      "properties": {
        "group_key": {
          "type": "string",
          "x-go-name": "GroupKey"
        },
        "lang": {
          "type": "string",
          "x-go-name": "Lang"
        },
        "lang_key": {
          "type": "string",
          "x-go-name": "LangKey"
        },
        "project_key": {
          "type": "string",
          "x-go-name": "ProjectKey"
        },
        "save_mode": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "SaveMode"
        },
        "status": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Status"
        },
        "translate_text": {
          "type": "string",
          "x-go-name": "TranslateText"
        }
      },
      "x-go-package": "github.com/qudj/fly_api/models"
    },
    "TransLg": {
      "type": "object",
      "properties": {
        "create_time": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "CreateTime"
        },
        "group_key": {
          "type": "string",
          "x-go-name": "GroupKey"
        },
        "lang": {
          "type": "string",
          "x-go-name": "Lang"
        },
        "lang_key": {
          "type": "string",
          "x-go-name": "LangKey"
        },
        "op_id": {
          "type": "string",
          "x-go-name": "OpId"
        },
        "op_name": {
          "type": "string",
          "x-go-name": "OpName"
        },
        "project_key": {
          "type": "string",
          "x-go-name": "ProjectKey"
        },
        "status": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Status"
        },
        "translate_text": {
          "type": "string",
          "x-go-name": "TranslateText"
        },
        "update_time": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "UpdateTime"
        }
      },
      "x-go-package": "github.com/qudj/fly_lib/models/proto/fly_starling_serv"
    }
  }
}